src/redgrease/requirements.txt
redis
attrs
watchdog
ConfigArgParse
pyyaml
